refactor(codex-tests): split auto-update suites
